+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+next
+----
+#. Up minimum jmbo-foundry to 0.5.
+#. Create a trivial south migration so order of migrations is correct.
+
 0.3.2
 -----
 #. Dev buildout now uses git instead of https.